Moved documentation generator exception to own package
