+++ detekt-generator/src/main/kotlin/io/gitlab/arturbosch/detekt/generator/collection/exception/InvalidCodeExampleDocumentationException.kt
@@ -1,4 +1,4 @@
-package io.gitlab.arturbosch.detekt.generator.collection
+package io.gitlab.arturbosch.detekt.generator.collection.exception
 
 class InvalidCodeExampleDocumentationException(ruleName: String)
 	: RuntimeException("Invalid rule documentation for noncompliant and compliant code examples in $ruleName.")
