+++ detekt-generator/src/main/kotlin/io/gitlab/arturbosch/detekt/generator/collection/exception/InvalidDocumentationException.kt
@@ -1,4 +1,4 @@
-package io.gitlab.arturbosch.detekt.generator.collection
+package io.gitlab.arturbosch.detekt.generator.collection.exception
 
 /**
  * Thrown to indicate that rule documentation in KDoc is missing or invalid.
