+++ detekt-generator/src/main/kotlin/io/gitlab/arturbosch/detekt/generator/collection/exception/InvalidIssueDeclaration.kt
@@ -1,4 +1,4 @@
-package io.gitlab.arturbosch.detekt.generator.collection
+package io.gitlab.arturbosch.detekt.generator.collection.exception
 
 class InvalidIssueDeclaration(attributeName: String)
 	: RuntimeException("Invalid issue declaration attribute $attributeName.")
